shop_srv/inventory_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
protobuf
loguru
python-consul2
nacos-sdk-python